isolate choose schedule component
