+++ src/pages/TimeTablePage.jsx
@@ -1,10 +1,9 @@
-import "../styles/utilities.css";
 import React, { useState } from "react";
-import TabPanel from "../components/Common/TabPanel";
 import Sidebar from "../components/TimeTableScheduler/Sidebar";
-import CourseSelection from "../components/TimeTableScheduler/CourseSelection";
+import ChooseSchedule from "../components/TimeTableScheduler/ChooseSchedule";
+import ChooseCoursesModal from "../components/TimeTableScheduler/ChooseCoursesModal";
 import { timeTableStyles } from "../styles/TimeTable";
-import { Box, Grid, Tabs, Tab } from "@mui/material";
+import { Box, Grid } from "@mui/material";
 
 const TimeTablePage = () => {
     const classes = timeTableStyles();
@@ -12,36 +11,9 @@ const TimeTablePage = () => {
     const [chosen, setChosen] = useState([]);
     const [modalOpen, setModalOpen] = useState(false);
 
-    const createOptionsArray = () => {
-        let optionsArray = [];
-        for (let i = 0; i < chosen.length; i++) {
-            optionsArray.push({ uc: chosen[i], class: null });
-        }
-        return optionsArray;
-    };
-
-    const tabs = Array(10).fill({
-        label: "Option",
-        options: createOptionsArray(),
-    });
-
-    const [value, setValue] = useState(0);
-
-    const handleChange = (event, newValue) => {
-        setValue(newValue);
-    };
-
-    function a11yProps(index) {
-        return {
-            id: `tab-${index}`,
-            key: `tab-${index}`,
-            "aria-controls": `tabpanel-${index}`,
-        };
-    }
-
     return (
         <div className={classes.container}>
-            <CourseSelection
+            <ChooseCoursesModal
                 majorHook={[major, setMajor]}
                 chosenHook={[chosen, setChosen]}
                 openHook={[modalOpen, setModalOpen]}
@@ -50,27 +22,7 @@ const TimeTablePage = () => {
             <Box sx={{ flexGrow: 1, padding: "2rem" }}>
                 <Grid container direction="row" justifyContent="space-evenly" alignItems="stretch" spacing={4}>
                     <Grid item xs={9} sm={9} md={9} lg={9} xl={9}>
-                        <Box sx={{ borderBottom: 1, borderColor: "divider" }}>
-                            <Tabs
-                                variant="fullWidth"
-                                value={value}
-                                onChange={handleChange}
-                                aria-label="TTS Option Tabs"
-                            >
-                                {tabs.map((tab, tabIdx) => (
-                                    <Tab label={`${tab.label} ${tabIdx + 1}`} {...a11yProps(tabIdx)} />
-                                ))}
-                            </Tabs>
-                        </Box>
-                        {tabs.map((tab, tabIdx) => (
-                            <TabPanel value={value} index={tabIdx} key={`tabpanel-${tabIdx}`}>
-                                {tab.options.map((option, optionIdx) => (
-                                    <span key={`option-${tabIdx}-${optionIdx}`}>
-                                        {option.uc}: {option.class}
-                                    </span>
-                                ))}
-                            </TabPanel>
-                        ))}
+                        <ChooseSchedule chosen={chosen} />
                     </Grid>
 
                     <Grid item xs={3} sm={3} md={3} lg={3} xl={3}>
